Don't follow symlinks when deleting directories
The existing commons-io implementation follows symlinks to recursively
delete files and I can't see how that would be desired behaviour. It did create
issues on some of our build-agents.

The main issue with symlinks and commons-io seems to be that `FileUtils.deleteQuietly`
doesn't check whether the initial argument is a symlink.
Upon recursing into the intial directory, subdirectories
get deleted by `forceDelete`, which in turn calls
`deleteDirectory` which does check symlinks.
(I've created an issue with commons-io as well:
https://issues.apache.org/jira/browse/IO-522)

This change swaps out calls to commons-io's `FileUtils.deleteQuietly`
with a local `FileUtil.deleteQuietly` method that calls
`FileUtils.forceDelete` directly and wraps the IOException
to behave the same as `deleteQuietly`

FileUtilTest didn't contain tests for the method I changed so I rather wrote a
test in the place that was relevant to the specific issue (DirectoryCleanerTest)

(Will likely fix https://github.com/gocd/gocd/issues/714 as well)